refactor: Separate function name classes from IBinaryLoader file
