+++ src/CoreHook.BinaryInjection/Loader/FunctionName.cs
@@ -1,0 +1,9 @@
+﻿
+namespace CoreHook.BinaryInjection.Loader
+{
+    public class FunctionName : IFunctionName
+    {
+        public string Module { get; set; }
+        public string Function { get; set; }
+    }
+}
