+++ src/CoreHook.BinaryInjection/Loader/IFunctionName.cs
@@ -1,0 +1,9 @@
+﻿
+namespace CoreHook.BinaryInjection.Loader
+{
+    public interface IFunctionName
+    {
+        string Module { get; }
+        string Function { get; }
+    }
+}
