+++ src/CoreHook.BinaryInjection/Loader/IRemoteFunctionCall.cs
@@ -1,0 +1,10 @@
+﻿using CoreHook.BinaryInjection.Loader.Serializer;
+
+namespace CoreHook.BinaryInjection.Loader
+{
+    public interface IRemoteFunctionCall
+    {
+        IFunctionName FunctionName { get; }
+        IBinarySerializer Arguments { get; }
+    }
+}
