+++ src/CoreHook.BinaryInjection/Loader/IRemoteManagedFunctionCall.cs
@@ -1,0 +1,8 @@
+﻿
+namespace CoreHook.BinaryInjection.Loader
+{
+    public interface IRemoteManagedFunctionCall : IRemoteFunctionCall
+    {
+        IAssemblyDelegate ManagedFunctionDelegate { get; }
+    }
+}
