+++ src/CoreHook.BinaryInjection/Loader/RemoteManagedFunctionCall.cs
@@ -1,0 +1,8 @@
+﻿
+namespace CoreHook.BinaryInjection.Loader
+{
+    public class RemoteManagedFunctionCall : RemoteFunctionCall, IRemoteManagedFunctionCall
+    {
+        public IAssemblyDelegate ManagedFunctionDelegate { get; set; }
+    }
+}
